+++ packages/core/src/components/ConnectionLine/index.ts
@@ -74,11 +74,7 @@ const ConnectionLine = defineComponent({
 
       const fromHandle = (startHandleId ? handleBounds.find((d) => d.id === startHandleId) : handleBounds[0]) ?? null
       const fromPosition = fromHandle?.position || Position.Top
-      const { x: fromX, y: fromY } = getHandlePosition(
-        fromPosition,
-        { ...fromNode.value.dimensions, ...fromNode.value.computedPosition },
-        fromHandle,
-      )
+      const { x: fromX, y: fromY } = getHandlePosition(fromNode.value, fromHandle, fromPosition)
 
       let toHandle: HandleElement | null = null
       if (toNode.value && connectionEndHandle.value?.handleId) {
@@ -97,8 +93,7 @@ const ConnectionLine = defineComponent({
         }
       }
 
-      // we assume the target position is opposite to the source position
-      const toPosition = fromPosition ? oppositePosition[fromPosition] : null
+      const toPosition = connectionEndHandle.value?.position ?? (fromPosition ? oppositePosition[fromPosition] : null)
 
       if (!fromPosition || !toPosition) {
         return null