+++ packages/core/src/composables/useKeyPress.ts
@@ -1,11 +1,11 @@
 import type { MaybeRefOrGetter } from 'vue'
-import { ref, toValue, watch } from 'vue'
+import { onMounted, ref, toRef, toValue, watch } from 'vue'
 import type { KeyFilter, KeyPredicate } from '@vueuse/core'
 import { onKeyStroke, useEventListener } from '@vueuse/core'
-import { useWindow } from './useWindow'
 
 export interface UseKeyPressOptions {
   actInsideInputWithModifier?: MaybeRefOrGetter<boolean>
+  target?: MaybeRefOrGetter<EventTarget | null | undefined>
 }
 
 export function isInputDOMNode(event: KeyboardEvent): boolean {
@@ -69,20 +69,16 @@ function useKeyOrCode(code: string, keysToWatch: string | string[]) {
 }
 
 /**
- * Reactive key press state
+ * Composable that returns a boolean value if a key is pressed
  *
- * todo: make this public?
- * @internal
- * @param keyFilter - Can be a boolean, a string or an array of strings. If it's a boolean, it will always return that value. If it's a string, it will return true if the key is pressed. If it's an array of strings, it will return true if any of the keys are pressed, or a combination is pressed (e.g. ['ctrl+a', 'ctrl+b'])
- * @param onChange - Callback function that will be called when the key state changes
+ * @public
+ * @param keyFilter - Can be a boolean, a string, an array of strings or a function that returns a boolean. If it's a boolean, it will act as if the key is always pressed. If it's a string, it will return true if a key matching that string is pressed. If it's an array of strings, it will return true if any of the strings match a key being pressed, or a combination (e.g. ['ctrl+a', 'ctrl+b'])
  * @param options - Options object
  */
-export function useKeyPress(
-  keyFilter: MaybeRefOrGetter<KeyFilter | null>,
-  onChange?: (keyPressed: boolean) => void,
-  options: UseKeyPressOptions = { actInsideInputWithModifier: true },
-) {
-  const window = useWindow()
+export function useKeyPress(keyFilter: MaybeRefOrGetter<KeyFilter | null>, options?: UseKeyPressOptions) {
+  const actInsideInputWithModifier = toRef(() => toValue(options?.actInsideInputWithModifier) ?? false)
+
+  const target = toRef(() => toValue(options?.target) ?? window)
 
   const isPressed = ref(toValue(keyFilter) === true)
 
@@ -91,12 +87,6 @@ export function useKeyPress(
   const pressedKeys = new Set<string>()
 
   let currentFilter = createKeyFilterFn(toValue(keyFilter))
-
-  watch(isPressed, (isKeyPressed, wasPressed) => {
-    if (isKeyPressed !== wasPressed) {
-      onChange?.(isKeyPressed)
-    }
-  })
 
   watch(
     () => toValue(keyFilter),
@@ -113,10 +103,8 @@ export function useKeyPress(
     },
   )
 
-  useEventListener(window, 'blur', () => {
-    if (toValue(keyFilter) !== true) {
-      isPressed.value = false
-    }
+  onMounted(() => {
+    useEventListener(window, ['blur', 'contextmenu'], reset)
   })
 
   onKeyStroke(
@@ -124,7 +112,7 @@ export function useKeyPress(
     (e) => {
       modifierPressed = wasModifierPressed(e)
 
-      const preventAction = (!modifierPressed || (modifierPressed && !options.actInsideInputWithModifier)) && isInputDOMNode(e)
+      const preventAction = (!modifierPressed || (modifierPressed && !actInsideInputWithModifier.value)) && isInputDOMNode(e)
 
       if (preventAction) {
         return
@@ -134,14 +122,14 @@ export function useKeyPress(
 
       isPressed.value = true
     },
-    { eventName: 'keydown' },
+    { eventName: 'keydown', target },
   )
 
   onKeyStroke(
     (...args) => currentFilter(...args),
     (e) => {
       if (isPressed.value) {
-        const preventAction = (!modifierPressed || (modifierPressed && !options.actInsideInputWithModifier)) && isInputDOMNode(e)
+        const preventAction = (!modifierPressed || (modifierPressed && !actInsideInputWithModifier.value)) && isInputDOMNode(e)
 
         if (preventAction) {
           return
@@ -150,10 +138,8 @@ export function useKeyPress(
         reset()
       }
     },
-    { eventName: 'keyup' },
+    { eventName: 'keyup', target },
   )
-
-  return isPressed
 
   function reset() {
     modifierPressed = false
@@ -184,4 +170,6 @@ export function useKeyPress(
 
     return keyFilter
   }
+
+  return isPressed
 }